+++ search-films/src/layouts/Header/index.jsx
@@ -5,15 +5,16 @@ import styles from './index.module.css';
  * Layout Header component 
  * @param {boolean} isAuthenticated - authenticated user or not
  * @param {string} userName - (byDefault=null) name of the user
+ * @param {callback} logout - (byDefault=null) can be used to log out
  * @returns {component} Header 
  * 
 */
-function Header({isAuthenticated=false, userName=null}) {
+function Header({isAuthenticated=false, userName=null, logout=null}) {
 	return (
 		<header className={styles.header}>
 			<div className={cs(styles['header-content'], 'restrict-content-size')}>
 				<Logo/>
-				<Menu isAuthenticated={isAuthenticated} userName={userName}/>
+				<Menu isAuthenticated={isAuthenticated} userName={userName} logout={logout}/>
 			</div>
 		</header>
 	);